pyproject.toml
[tool.poetry]
name = "redditbot"
version = "0.1.0"
description = ""
authors = ["Leonpei <peileon2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
praw = "^7.8.1"
openai = "^1.73.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
